+++ pom.xml
@@ -4,52 +4,77 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.aman.flink</groupId>
-    <artifactId>couchbase-starter</artifactId>
+    <groupId>com.github.isopropylcyanide</groupId>
+    <artifactId>flink-couchbase-sink</artifactId>
     <version>1.0.0</version>
+
+    <name>Flink Couchbase Sink</name>
+    <description>Provides a Couchbase sink for dumping documents using Flink</description>
+    <url>https://github.com/isopropylcyanide/flink-couchbase-data-starter</url>
+
+    <developers>
+        <developer>
+            <name>Aman Garg</name>
+            <email>amangargcse@outlook.com</email>
+            <timezone>5</timezone>
+            <id>isopropylcyanide</id>
+        </developer>
+    </developers>
+
+    <licenses>
+        <license>
+            <name>Apache 2.0</name>
+            <url>http://www.apache.org/licenses/LICENSE-2.0.txt</url>
+            <distribution>repo</distribution>
+        </license>
+    </licenses>
+
+    <properties>
+        <!--Build-->
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
+        <maven.compiler.source>1.8</maven.compiler.source>
+        <maven.compiler.target>1.8</maven.compiler.target>
+
+        <!--Application-->
+        <flink.java.version>1.6.2</flink.java.version>
+        <flink.streaming.java_2.11.version>1.6.2</flink.streaming.java_2.11.version>
+        <flink.clients_2.11.version>1.6.2</flink.clients_2.11.version>
+        <couchbase.java.client.version>2.6.2</couchbase.java.client.version>
+    </properties>
 
     <dependencies>
         <dependency>
             <groupId>org.apache.flink</groupId>
             <artifactId>flink-java</artifactId>
-            <version>1.6.2</version>
+            <version>${flink.java.version}</version>
         </dependency>
+
         <dependency>
             <groupId>org.apache.flink</groupId>
             <artifactId>flink-streaming-java_2.11</artifactId>
-            <version>1.6.2</version>
+            <version>${flink.streaming.java_2.11.version}</version>
         </dependency>
+
         <dependency>
             <groupId>org.apache.flink</groupId>
             <artifactId>flink-clients_2.11</artifactId>
-            <version>1.6.2</version>
+            <version>${flink.clients_2.11.version}</version>
         </dependency>
 
-        <!--Couchbase dependency -->
         <dependency>
             <groupId>com.couchbase.client</groupId>
             <artifactId>java-client</artifactId>
-            <version>2.6.2</version>
+            <version>${couchbase.java.client.version}</version>
         </dependency>
-
     </dependencies>
 
     <build>
         <plugins>
-            <!--Java 8 compiler/plugin-->
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <configuration>
-                    <source>8</source>
-                    <target>8</target>
-                </configuration>
-            </plugin>
-
             <!--Maven shade plugin for fat jar. Alternatively, put libs in %flink%/lib-->
             <!-- We use the maven-shade plugin to create a fat jar that contains all dependencies
            except flink and it's transitive dependencies. The resulting fat-jar can be executed
-           on a cluster. Change the value of Program-Class if your program entry point changes. -->
+           on a cluster. Change the value of Main-Class if your program's entry point changes. -->
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-shade-plugin</artifactId>
@@ -66,7 +91,7 @@
                                 <!-- add Main-Class to manifest file -->
                                 <transformer
                                         implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>com.aman.flink.job.FlinkDatabaseStartupJob</mainClass>
+                                    <mainClass>com.github.isopropylcyanide.flinkcouchbasesink.job.FlinkDatabaseStartupJob</mainClass>
                                 </transformer>
                             </transformers>
                             <createDependencyReducedPom>false</createDependencyReducedPom>
