+++ src/main/java/com/github/isopropylcyanide/flinkcouchbasesink/job/FlinkDatabaseStartupJob.java
@@ -1,11 +1,24 @@
-package com.aman.flink.job;
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");	 * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.	 * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at	 * You may obtain a copy of the License at
+ *	 *
+ * http://www.apache.org/licenses/LICENSE-2.0	 * http://www.apache.org/licenses/LICENSE-2.0
+ *	 *
+ * Unless required by applicable law or agreed to in writing, software	 * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,	 * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.	 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and	 * See the License for the specific language governing permissions and
+ * limitations under the License.	 * limitations under the License.
+ */
+package com.github.isopropylcyanide.flinkcouchbasesink.job;
 
-import com.aman.flink.Env;
-import com.aman.flink.ResponseSubscriber;
-import com.aman.flink.constants.Constant;
-import com.aman.flink.model.StarterJsonDocument;
-import com.aman.flink.utility.CouchbaseManager;
-import com.aman.flink.utility.Util;
+import com.github.isopropylcyanide.flinkcouchbasesink.Env;
+import com.github.isopropylcyanide.flinkcouchbasesink.ResponseSubscriber;
+import com.github.isopropylcyanide.flinkcouchbasesink.constants.Constant;
+import com.github.isopropylcyanide.flinkcouchbasesink.model.StarterJsonDocument;
+import com.github.isopropylcyanide.flinkcouchbasesink.utility.CouchbaseManager;
+import com.github.isopropylcyanide.flinkcouchbasesink.utility.Util;
 import com.couchbase.client.java.document.json.JsonObject;
 import org.apache.flink.api.java.io.TextInputFormat;
 import org.apache.flink.core.fs.Path;
