+++ src/main/java/com/github/isopropylcyanide/flinkcouchbasesink/utility/Util.java
@@ -1,8 +1,21 @@
-package com.aman.flink.utility;
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");	 * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.	 * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at	 * You may obtain a copy of the License at
+ *	 *
+ * http://www.apache.org/licenses/LICENSE-2.0	 * http://www.apache.org/licenses/LICENSE-2.0
+ *	 *
+ * Unless required by applicable law or agreed to in writing, software	 * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,	 * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.	 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and	 * See the License for the specific language governing permissions and
+ * limitations under the License.	 * limitations under the License.
+ */
+package com.github.isopropylcyanide.flinkcouchbasesink.utility;
 
-import com.aman.flink.Env;
-import com.aman.flink.constants.Constant;
-import com.aman.flink.model.StarterJsonDocument;
+import com.github.isopropylcyanide.flinkcouchbasesink.Env;
+import com.github.isopropylcyanide.flinkcouchbasesink.constants.Constant;
+import com.github.isopropylcyanide.flinkcouchbasesink.model.StarterJsonDocument;
 import com.couchbase.client.deps.com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.flink.api.java.utils.ParameterTool;
 import org.slf4j.Logger;
