+++ src/main/resources/config.properties
@@ -1,12 +1,17 @@
 # Couchbase node
 couchbase.node=localhost
+
 # Couchbase username
 couchbase.username=aman
+
 # Couchbase Password
 couchbase.password=hellohi123
+
 # Startup document script
 startup.documents.path=C:\\Users\\Acco-Blr-01\\Desktop\\couchbasestarter\\src\\main\\resources\\startup-documents.json
+
 # Startup document polling behaviour
 startup.documents.poll.continuous=false
+
 # Startup document polling duration if behaviour is set
 startup.documents.poll.duration=30000
